+++ test/dummy/config/environments/test.rb
@@ -5,5 +5,6 @@ Rails.application.configure do
   config.cache_store = :memory_store
   config.active_support.deprecation = :stderr
   config.active_job.queue_adapter = :test
+  config.action_controller.allow_forgery_protection = false
   config.active_record.encryption.encrypt_fixtures = true
 end